feat: add merge visited urls files
