+++ example-files/urls-0.json
@@ -1,0 +1,9 @@
+{
+  "cypress/e2e/homepage.cy.js": {
+    "The home page / goes using the link": ["/public/index.html"]
+  },
+  "cypress/e2e/utils.cy.js": {
+    "updateVisitedUrls / sorts the spec names": [],
+    "updateVisitedUrls / sorts the test titles for each spec": []
+  }
+}
